Add option for changing the name of the css output file (default fonts.css).
This adds an additional option wherein one can configure the name of the css output file (before it was a hard-coded "fonts.css"). Although one could pipe the output through gulp-rename or similar, this allows for keeping one's gulp file cleaner. This can be useful for different setups (such as outputting as _fonts.scss for use with a sass task).